[bfcache] Report null when iframes' atrributes are null
In HTMLIframeElement::ConvertToReportValue(), we used to convert null to
an empty string. This function builds a value based on iframes'
attributes when they change. They are sent to the browser side.

The only web-exposed usage of this is NotRestoredReasons API.
NotRestoredReasons used to report an empty string for non-set iframes'
attributes.

With this fix, NRR API starts reporting null for non-set attributes.
No tests broke with this change, except for the NotRestoredReasons tests
that expected the wrong thing (an empty string instead of null).

Bug: 331883460
Change-Id: Ic15d75562e7f473c34b8c85aea6900ff8f51d961
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5394900
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: Yuzu Saijo <yuzus@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280111}