[2/n] Fix TabDragging tests for fallback tab dragging: CancelOnNewTab
This is part of a series of CLs to make the TabDragging tests pass with
fallback tab dragging, i.e. with the AllowWindowDragUsingSystemDragDrop
feature enabled. The goal is for all tests that pass with the current
tab dragging implementation to pass with fallback tab dragging as well,
so that it can be enabled by default.

This CL changes the CancelOnNewTab test to be skipped when fallback tab
dragging will be used, as it doesn't make sense in that case: a new tab
can only be added to the drag source window as there is no dragged
window, which isn't problematic, so we don't need to cancel the drag.

Bug: 1361940
Change-Id: I60b62945268dfc847b56a41997840a2acbc42d4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5253165
Commit-Queue: Max Ihlenfeldt <max@igalia.com>
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280135}