graphite: Remove separate ANGLEDebugLayer feature flag
Currently, SkiaGraphite feature is checked in the browser process after
making an educated guess about whether it will be possible for the GPU
process to use Graphite.

On Mac, this involves checking ANGLE Metal feature state and the
hardware model name against a list to match ANGLE Metal's blocklist.
However, based on UMA data from the wild, we see cases where ANGLE
Metal is disabled, but we're still registered in the enabled group for
Graphite - this is precisely the scenario the early check was designed
to avoid.

Closer inspection of the experiment config by vasilyt@ points to
ANGLEDebugLayer feature check as a culprit. This feature is configured
in SkiaGraphite experiment groups and checked in gl_display.cc code
independently of SkiaGraphite feature state. Checking a feature can
enroll the client in experiment groups that specify that feature.
Therefore, it's possible that a client is falling into the SkiaGraphite
enabled arm only because the client checked ANGLEDebugLayer.

This CL fixes that by only enabling ANGLE's debug layer if GrContextType
is GraphiteDawn and Graphite backend validation FeatureParam is enabled,
getting rid of the separate ANGLEDebugLayer feature flag as a result. It
also removes the SkiaGraphiteDawnShareDevice FeatureParam that's never
set to a non-default value via finch and is obsolete at this point.

Bug: 331680185
Change-Id: I470846e311e89e63f793783b52ee01d46e85e17e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5403504
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Auto-Submit: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280143}