[iOS] Cleanup the temporary additional opt in for Bookmarks and RL
In the first launch of butter for bookmarks on iOS; Bookmarks and RL
required an additional opt-in to be enabled.
This is no longer needed after kReplaceSyncPromosWithSignInPromos was
enabled by default.

This CL cleans up all the references to the dedicated pref for this
additional opt-in, expect for the migration logic, which will be removed
later.

Fixed: 40266203
Change-Id: Ic6516691af41e91f4626028bba2af0a3e1c23b5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5401722
Auto-Submit: Jood Hajeer <jood@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280150}