[PWA] Remove OS Hooks Errors from OS integration manager
To further break the cleanup of the non synchronize OS integration
flow, this CL removes the OsHooksErrors struct. Most of the changes in the CL is mechanical, and involves removing instances to it and cleaning up the os_integration_manager include.

The main changes to support the new Synchronize() behavior are:
1. `RemoveWebAppJob`: Instead of relying on the errors to determine
success rate, a web app removal should wipe out the entire
os_integration_state field in the web_app DB. This CL introduces
behavior that measures the before and after state to determine if
a removal of OS states were successful.
2. To support 1, minor changes were made into
`OsIntegrationManager::WriteStateToDB()`, which would not write
the state if web_app->is_uninstalling was set to true. That condition
was removed.
3. Additionally, the `os_uninstallation_registration` field would
never be empty once set, even during uninstallation, so this CL takes
care of that by emptying that proto inside the OS integration state.

For manifest updates, the kUpdateTaskFailed is not used anymore
starting this CL, so that will be removed in a subsequent CL.

Bug: 328226797
Change-Id: Ib674e19d9817f125bb6f5b34dd1a4ad31749a865
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5373844
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280161}