Implement ApplyIncrementalSyncChanges in ProductSpecificationsSyncBridge
ProductSpecificationsSyncBridge receives changes from the sync server
via ApplyIncrementalSyncChanges. The local store and in-memory
representations of the specifics are updated. Where there is a
collision, the latest specifics/product specifications are used.
MergeFullSyncData is the same as ApplyIncrementalSyncChanges because
the user cannot access product specifications unless sync is turned on
(i.e. there is no local data to merge).

Bug: b:329519488, b:329519487
Change-Id: I95d1ef1d54f08c94d12f1f2a45fefd03e7c2ac92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5406304
Commit-Queue: David Maunder <davidjm@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280212}