[FPF] Create new fingerprinting_protection_filter in components/.
A new component is needed because the code will be shared between Blink and the browser process.
In order to have comprehensive filtering of all resource loads we will need filtering logic both
within the browser and on renderer processes. The renderer-side filtering will also depend on
Page-level activation in and communication from the browser.

It's a different component from components/subresource_filter/ because it caters to a new
ruleset anddifferent activation conditions but the directory structure will be similar.


Bug: 331220706
Change-Id: Ic85e8eb91164fb08e3cea1c5105e2611bfd2ec60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5392546
Commit-Queue: Zainab Rizvi <rizvis@google.com>
Reviewed-by: Patricia Alfonso <trishalfonso@google.com>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280213}