[Read Aloud] Ensure Read Aloud works with PDFs
Previously, Read Aloud checked for a starting node by checking that
the first text node id was not equal to -1. But this was an incorrect
assumption, as PDF node ids may be negative.

This CL cleans up the Read Aloud code so that we're now using null to
track whether or not a node id has been set, instead of using -1.

Bug: 329109390
Change-Id: Ie2087f5b367618498760bdf3f869624479cc4a2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5405797
Commit-Queue: Lauren Winston <lwinston@google.com>
Reviewed-by: Jocelyn Tran <jocelyntran@google.com>
Cr-Commit-Position: refs/heads/main@{#1280237}