[CrOS VPN] Only check if there is an always-on vpn package name
Instead of checking whether the always-on vpn package has a true
lock-down value, check if there is a non-empty name. This is
because the lock-down value is true right when the admin activates
an always-on vpn package, but turns false shortly after. The
vpn package name is always non-empty when the admin specifies
an always-on vpn, and empty when the admin does not specify
an always-on vpn.

Bug: b/302744527
Test: browser_test, chromeos_unittest, and manual
Change-Id: I271953e2541b27bc77a6cc05c411fb3e6777fd2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5372902
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Reviewed-by: Jason Zhang <jiajunz@google.com>
Cr-Commit-Position: refs/heads/main@{#1280265}