Add ancestor chain to partition key and update call sites and tests
This is the initial implementation of the ancestor chain bit addition to the cookie partition key.

The unit tests associated with the partition key have also been added and/or updated accommodate the new value.

For all the methods with parameters added to them, they have been given default values so that preexisting code will be able to work with them. This was done to allow for the wide spread changes to the codebase to be broken up into smaller CLs and after all those land another CL to address https://crbug.com/325439525, will be used to change the methods to require values be passed.

Bug: 41486025
Bug: 325439525
Change-Id: I9f0fec7dc4bee44031c5e80648e451d2fa733628
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5296867
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Reviewed-by: Dylan Cutler <dylancutler@google.com>
Commit-Queue: Aaron Selya <selya@google.com>
Cr-Commit-Position: refs/heads/main@{#1280278}