[Read Aloud] Remove reference to fontTemplate.render()
In attempting to refactor menus to use cr-lazy-render, I needed to get
rid of fontTemplate.render(), but it turns out that a lot of
our tests (even outside of tests for font menu) were dependent on
this line to render everything properly.

I think this line was essentially obscuring that we need calls to
flush() in tests to force re-rendering.

Everything renders as expected in the toolbar with and without the
Read Aloud flag in an actual build of Chrome.

Bug: 329676284
Change-Id: I8ede75043c64a48b5cf7f61be06d8feea7528a9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5407357
Reviewed-by: Eitan Goldberger <eitang@google.com>
Commit-Queue: Lauren Winston <lwinston@google.com>
Cr-Commit-Position: refs/heads/main@{#1280319}