[Fix It] Wait to refresh private certificates until adapter is ready
On BlueZ, waits until the adapter is retrieved from the
BlueotothAdapterFactory before refreshing private certificates. On
Floss, waits until the adapter is retrieved and powered on before
refreshing private certificates. This is necessary because the
Adapter's address is required to refresh the private certificates,
and on Floss, if the adapter is powered off, the address is not
available.

back on.

Tested: Manaully verified NS on Floss and BlueZ. Triggered private certificates regeneration on Floss when the adapter is off, and observed that the certificates were not refreshed until the adapter was powered
Bug: b/315277593
Change-Id: If7cb45e471063ce55ac1bc9a1f51a6251857e7b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5402460
Commit-Queue: Juliet Lévesque <julietlevesque@google.com>
Reviewed-by: Jack Shira <jackshira@google.com>
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280412}