[AriaNotify] Fire UIA notification events for ARIA notifications
In Windows, we want to fire an UIA notification event for each posted
ARIA notification using the `UiaRaiseNotificationEvent` method.

This method expects a `NotificationProcessing` enum that we map from the
`interrupt` and `priority` options provided in the `ariaNotify` call.
Also, the notification id string is provided to UIA via the `activityId`
parameter, which will enable filtering capabilities for consumers.

Bug: 326277796
Change-Id: I46372c25280f22ab16ec3db1fc243bfc5a53476a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5399678
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Ethan Jimenez <ethavar@microsoft.com>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1280433}