[Fix It] Record frequency of eligible device list unexpectedly empty
Adds a metric to MultiDeviceSetupImpl::GetEligibleActiveHostDevices()
to log the number of elgible device hosts. This will help us understand the frequency of the eligible host device list being
empty. This is unexpected because to get to the point in the
Multidevice Setup flow when
MultiDeviceSetupImpl::GetEligibleActiveHostDevices() is called by the
Javascript layer, it is expected that there is an eligible host that is
used to calculate a supported host status, which is then used to
determine whether the setup flow should beign or not. This metric
will help us understand the severity of how often users are in this
bad state.

metric is emmitted to during setup.

Tested: Manually verified on DUT using chrome:://histograms that this
Bug: b/320789583
Change-Id: Iff2c1b7e2e4804256a8586a7c801308fc3518b5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5408151
Reviewed-by: Jon Mann <jonmann@chromium.org>
Commit-Queue: Juliet Lévesque <julietlevesque@google.com>
Cr-Commit-Position: refs/heads/main@{#1280436}