[remoting] Use SequenceChecker in DesktopSessionProxy.
This removes `caller_task_runner_` and replaces it with a
SequenceChecker, with some member guards. The task-runner was used for
deleting the ref-counted DesktopSessionProxy on the caller thread. This
CL uses RefCountedDeleteOnSequence instead, so `caller_task_runner_` is
no longer needed.

All DesktopSessionProxy methods are documented to run on the caller
thread, with some exceptions. The only exceptions noted were for
audio-capture, so there are no new guards or checks around that.

This also adds some new checks for methods that are expected to run on
the caller thread:
CreateFileOperations()
GetCapabilities()
SetCapabilities()
OnDesktopSessionAgentStarted()

Change-Id: I392e4743fb23a138a8cd3daba690bc8d22feaf48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5405906
Auto-Submit: Lambros Lambrou <lambroslambrou@chromium.org>
Reviewed-by: Joe Downing <joedow@chromium.org>
Commit-Queue: Lambros Lambrou <lambroslambrou@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280438}