bluetooth: DeviceChanged for classic devices on Mac
BluetoothAdapterMac periodically polls the platform
Bluetooth API to check for paired Bluetooth Classic devices.
Previously, observers would be notified when a new device
was added but would not be notified for any device
changes. With this CL, observers are notified when the UUID
set changes.

This CL also adds chrome://device-log logging for events
related to this polling behavior.

Bug: 329866142
Change-Id: I9a81d9e387e0d45830775a688ac47889c6861a63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5406216
Commit-Queue: Matt Reynolds <mattreynolds@chromium.org>
Reviewed-by: Alvin Ji <alvinji@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280445}