Fix HttpStreamParser::current_response_start_time_ in the case of a 0-length response body.
HttpStreamParser was failing to set current_response_start_time_ in
the case of HEAD requests and responses with a "Content-Length: 0"
header.  This CL fixes that issue.

We still fail to set it in the case we receive a partial request
header with no body over HTTP (Not SSL, which we consider a fatal
error), but that's probably not worth fixing.

Fixed: 331834510
Change-Id: Id4bcbe6bbc560de98f475c3906d76773cfc7cad0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5405424
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Commit-Queue: mmenke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280475}