Permit truncating entries with full memory cache
By creating entries with no data it is possible to exceed the permitted
size of the memory cache. This happens when the fuzzer is run against
the cache. Since reducing the size of a cache that is oversized is not
harmful, remove the CHECK statements that assume the cache is not
oversized.

BUG=331839344

Change-Id: Iabed5e9403233703f17787331f942197b5ba7f81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5401315
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280657}