Port a test to Lacros
BrowserNonClientFrameViewChromeOSTouchTestWithWebUiTabStrip's
TabletSplitViewSwipeDownFromEdgeOpensWebUiTabStrip test sent a touch
event to a point outside the browser window. In the case of Lacros,
that would require using ui_controls (and making the test run as part
of interactive_ui_tests). Fortunately, we can just as well use a point
inside the browser window for the purpose of the test.

Bug: 1494785
Change-Id: Ic6475e5167b7ce84a6dcbb2342e70689bcab8d8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5409792
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Georg Neis <neis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280670}