[Extensions] Confirm that worker stop notification is on UI thread.
As seen in the bug, //extensions doesn't always get the worker thread
termination (WorkerThread::PrepareForShutdownOnWorkerThread())
notification (ServiceWorkerHost::DidStopServiceWorkerContext()) before ServiceWorkerVersion::OnTimeoutTimer() is called for the shutdown
timing out. OnTimeoutTimer() causes ~ServiceWorkerHost for the previous
worker so even if the notification is eventually sent it won't be seen.

This causes //extensions to not clean up a previous record
of the worker in WorkerIdSet There are also other situations where
synchronous notification of worker stop would help.

//extensions would like to make this notification synchronous to help
with this.

From my own local testing, it seems that all the //content browser
process logic for the running status notifications are on the same
thread as the //extensions logic. This was likely due to the
ServiceWorkerOnUI project. I'd like to DCHECK this to confirm before
assuming.

Bug: 40936639
Change-Id: Id019124159abe5bc68c6d2822083f14fb8c509a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5405658
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Auto-Submit: Justin Lulejian <jlulejian@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280671}