[//gpu] Initialize SICaps' MacOS-specific texture target to 0
As with the other fields, the creator of a SharedImageCapabilities
instance must ensure that this field is set to an appropriate value.
This CL initializes it to 0 to ensure that it has a reliable sentinel
value for easing debugging of cases where the creator misses setting
the value.

Bug: 41494843
Change-Id: I9ae795eaecd8dbe0177536f9e008a8c50906bcb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5402274
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Quyen Le <lehoangquyen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280690}