[//gpu] Add unittest of ClientSI creation from SharedImageInterface
This CL adds a unittest of ClientSharedImage creation from
SharedImageInterface. To do so, we add a TestSharedImageInterface
implementation in client_shared_image_unittest.cc (the only existing one
is in //components/viz and is much heavierweight). We stub everything
out except for the one function needed to create a ClientSI in the
initial unittest. We will fill it out as the tests expand over time.

Note: While it might seem odd to create ClientSharedImage from
SharedImageInterface in the ClientSharedImage unittest, this is the
only and intended supported model of creating ClientSharedImages beyond
ImportUnowned() (which is already tested). Unless going through
ImportUnowned(), a ClientSharedImage instance *must* be supplied a
SharedImageInterfaceHolder, and that can *only* be given by a subclass
of SharedImageInterface.

Bug: 40286368
Change-Id: Ia47ed9250c7be46a1508e8310b32d53c37245857
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5407653
Reviewed-by: Quyen Le <lehoangquyen@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1280691}